+++ Mediateka.Domain/Mediateka.Domain.csproj
