+++ Mediateka.Domain/Model/Artist.cs
