+++ Mediateka.Domain/Services/IAlbumRepository.cs
@@ -1,13 +1,15 @@
-﻿namespace Mediateka.Application.Contracts;
+﻿using Mediateka.Domain.Model;
+namespace Mediateka.Domain.Services;
+
 /// <summary>
-/// Интерфейс для службы, выполняющей аналитические запросы согласно бизнес-логике приложения
+/// Наследник обобщенного интерфейса для альбомов с дополнительной функциональностью 
 /// </summary>
-public interface IAlbumAnalyticsService
+public interface IAlbumRepository : IRepository<Album, int>
 {
+
     /// <summary>
     /// Возвращает список треков для указанного альбома
     /// </summary>
-    /// <param name="albumId">ID альбома.</param>
     /// <returns>Список строк с информацией о треках.</returns>
     public IList<string> GetTreksInAlbum();
 
@@ -28,16 +30,4 @@ public interface IAlbumAnalyticsService
     /// <returns>Список с минимальной, средней и максимальной длительностью альбомов.</returns>
     public Task<IList<(int MinDuration, double avgDuration, int maxDuration)>> GetAlbumDurationStatistics();
 
-}
-public interface IArtistAnalyticsService
-{
-    /// <summary>
-    /// Возвращает информацию о всех артистах в виде списка строк.
-    /// </summary>
-    public IList<string> GetAllArtisInfo();
-    /// <summary>
-    /// Метод для вывода артистов с максимальныи количествм альбомов
-    /// </summary>
-    /// <returns>Список артистов и альбомов</returns>
-    public IList<string> GetArtistsWithMaxAlbums();
 }
