+++ Mediateka.Domain/Services/IArtistRepository.cs
@@ -1,14 +1,17 @@
 ﻿using Mediateka.Domain.Model;
 namespace Mediateka.Domain.Services;
+
 /// <summary>
 /// Наследник обобщенного интерфейса для артистов с дополнительной функциональностью 
 /// </summary>
 public interface IArtistRepository : IRepository<Artist, int>
 {
+
     /// <summary>
     /// Возвращает информацию о всех артистах в виде списка строк.
     /// </summary>
     public IList<string> GetAllArtisInfo();
+
     /// <summary>
     /// Метод для вывода артистов с максимальныи количествм альбомов
     /// </summary>
