--- Mediateka/Mediateka.Application.Contracts/Genre/GenreCreateUpdateDto.cs
@@ -1,6 +1,0 @@
-﻿namespace Mediateka.Application.Contracts.Genre;
-/// <summary>
-/// Dto для создания или изменения жанра
-/// </summary>
-/// <param name="Name">Название жанра</param>
-public record GenreCreateUpdateDto(string? Name);
