--- Mediateka/Mediateka.Application.Contracts/Mediateka.Application.Contracts.csproj
@@ -1,13 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net8.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Mediateka.Domain\Mediateka.Domain.csproj" />
-  </ItemGroup>
-
-</Project>
