--- Mediateka/Mediateka.Application/Mediateka.Application.csproj
@@ -1,18 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net8.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="AutoMapper" Version="14.0.0" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Mediateka.Application.Contracts\Mediateka.Application.Contracts.csproj" />
-    <ProjectReference Include="..\Mediateka.Domain\Mediateka.Domain.csproj" />
-  </ItemGroup>
-
-</Project>
